feat(field type): move field-types into `types` folder.
