+++ src/templates/field.wrapper.ts
@@ -1,4 +1,4 @@
-import {Component, ViewChild, ViewContainerRef} from "@angular/core";
+import {ViewContainerRef} from "@angular/core";
 import {Field} from "./field";
 
 export abstract class FieldWrapper extends Field {
