--- src/templates/types.ts
@@ -1,15 +1,0 @@
-import {FormlyFieldCheckbox} from "./formlyfield.checkbox";
-import {FormlyFieldMultiCheckbox} from "./formlyfield.multicheckbox";
-import {FormlyFieldInput} from "./formlyfield.input";
-import {FormlyFieldRadio} from "./formlyfield.radio";
-import {FormlyFieldTextArea} from "./formlyfield.textarea";
-import {FormlyFieldSelect} from "./formlyfield.select";
-
-export {
-  FormlyFieldCheckbox,
-  FormlyFieldMultiCheckbox,
-  FormlyFieldInput,
-  FormlyFieldRadio,
-  FormlyFieldTextArea,
-  FormlyFieldSelect,
-};
