+++ src/templates/types/types.ts
@@ -1,0 +1,15 @@
+import {FormlyFieldCheckbox} from "./checkbox";
+import {FormlyFieldMultiCheckbox} from "./multicheckbox";
+import {FormlyFieldInput} from "./input";
+import {FormlyFieldRadio} from "./radio";
+import {FormlyFieldTextArea} from "./textarea";
+import {FormlyFieldSelect} from "./select";
+
+export {
+  FormlyFieldCheckbox,
+  FormlyFieldMultiCheckbox,
+  FormlyFieldInput,
+  FormlyFieldRadio,
+  FormlyFieldTextArea,
+  FormlyFieldSelect,
+};
